Update Au4 application

+++ lib/au4/context/unit.ex
@@ -6,7 +6,7 @@ defmodule Au4.Context.Unit do
     field :name, :string
     field :description, :string
     field :booking, :string
-    field :price, :decimal
+    field :price, :integer
 
     belongs_to :floor, Au4.Context.Floor
     many_to_many :user, Au4.Account.User, join_through: Au4.Context.UserApartment, on_replace: :delete
